requirements.txt
APScheduler==3.10.4
black==24.2.0
click==8.1.7
loguru==0.7.2
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
TgCrypto==1.2.5
tzlocal==5.2
uvloop==0.19.0
